Revert "Fix build due to ut64 ≡ uint64_t"
This reverts commit dfbcc6993c4f37d9d86e3bec7231581f4ed41193.

+++ src/widgets/BreakpointWidget.cpp
@@ -78,7 +78,7 @@ QVariant BreakpointModel::data(const QModelIndex &index, int role) const
     case Qt::EditRole:
         switch (index.column()) {
         case AddrColumn:
-            return (quint64)breakpoint.addr;
+            return breakpoint.addr;
         case TraceColumn:
             return breakpoint.trace;
         case EnabledColumn:
